% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_format_col.R
\name{dt_fmt_col}
\alias{dt_fmt_col}
\alias{fmt_dtcol}
\title{Colour DT column by range}
\usage{
dt_fmt_col(
  dt,
  columns,
  column_range = "auto",
  colours = c("#9162C5", "#F1F1F1", "#6BA359"),
  reverse_colours = FALSE
)
}
\arguments{
\item{dt}{a DT object created by \code{DT::datatable()}}

\item{columns}{a character or numeric vector of column identifiers (column names or indices)}

\item{column_range}{numeric or "auto": either the minimum and maximum range to
be coloured, or "auto" which defaults to the max and min of the table column.}

\item{colours}{A set of hex colours to be passed on to \code{colorRampPalette()},
defaults to purple-green}

\item{reverse_colours}{logical: reverses direction of colour scale}
}
\value{
A DT object with added colour formatting
}
\description{
Colours DT columns based on the minimum and maximum of each column.
Defaults to a purple-green colour scheme HULK which is colourblind-friendly,
and naturally associates "green" to "good" and "purple" to "bad".
}
\examples{
DT::datatable(mtcars) |>
  # standard: formats high values as green, low values as purple
  dt_fmt_col(columns = c("hp", "cyl")) |>
  # reverse: formats low values as green, high values as purple
  dt_fmt_col(columns = "mpg", reverse_colours = TRUE) |>
  # custom colours
  dt_fmt_col(columns = "wt", colours = c("orange", "white", "blue"))

}
