% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_imports.R
\name{fontimport_plex_sans}
\alias{fontimport_plex_sans}
\alias{fontimport_plex_condensed}
\alias{fontimport_bai_jamjuree}
\alias{fontimport_jost}
\title{font imports}
\usage{
fontimport_plex_sans()

fontimport_plex_condensed()

fontimport_bai_jamjuree()

fontimport_jost()
}
\description{
font imports

font imports

font imports

font imports
}
\keyword{internal}
