% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_inputs.R
\name{gen_input_map}
\alias{gen_input_map}
\title{Generate Shiny Inputs for a vector of identifiers}
\usage{
gen_input_map(uid, FUN, id_prefix = NULL, ...)
}
\arguments{
\item{uid}{a vector of unique identifiers}

\item{FUN}{a ShinyInput function - should theoretically work for all shinyInput functions that have \code{inputid} as the first argument}

\item{id_prefix}{a string that will be combined with the unique identifier and passed as the inputid}

\item{...}{other arguments to be passed to the \code{FUN}}
}
\value{
a character vector of shinyInputs
}
\description{
Generate Shiny Inputs for a vector of identifiers
}
\examples{
gen_input_map(1:5, shiny::numericInput, id_prefix = "playerid_", label = "my_label", value = 1)
gen_input_map(FUN = numericInput, id_prefix = "itemprice_", uid = c(1, 3, 5, 7), ...)
}
