% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_inputs.R
\name{gen_input_seq}
\alias{gen_input_seq}
\title{Generate a sequence of Shiny Inputs}
\usage{
gen_input_seq(FUN, id_prefix, len, ...)
}
\arguments{
\item{FUN}{a ShinyInput function - should theoretically work for all shinyInput functions that have \code{inputid} as the first argument}

\item{id_prefix}{a string that will be combined with a sequence of length \code{len} and passed as the inputid}

\item{len}{the desired output length - generates a sequence that is combined with\code{id_prefix} and passed as the inputid}

\item{...}{other arguments to be passed to the \code{FUN}}
}
\value{
a character vector of shinyInputs
}
\description{
Generate a sequence of Shiny Inputs
}
\examples{
gen_input_seq(FUN = numericInput, id_prefix = "itemprice_", len = 5, ...)
}
