% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uv.R
\name{set_geom_colour_defaults}
\alias{set_geom_colour_defaults}
\title{Update geom defaults}
\usage{
set_geom_colour_defaults(colour = "#57c1f1")
}
\arguments{
\item{colour}{colour of geom default}
}
\description{
Update geom defaults
}
