% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tantastic.R
\name{set_geom_defaults}
\alias{set_geom_defaults}
\title{Update geom defaults}
\usage{
set_geom_defaults(colour)
}
\arguments{
\item{colour}{colour of geom default}
}
\description{
Update geom defaults
}
