% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uv.R
\name{set_geom_font_defaults}
\alias{set_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
set_geom_font_defaults(
  family = "IBM Plex Sans",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)
}
\arguments{
\item{family, face, size, color}{font family name, face, size and color}
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label} and \link[ggplot2:geom_text]{ggplot2::geom_text} font defaults
}
