% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_timezone.R
\name{use_client_tz}
\alias{use_client_tz}
\alias{get_client_tz}
\title{Get User's Reported Time Zone}
\usage{
use_client_tz(inputId = "_client_tz")

get_client_tz(
  inputId = "_client_tz",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{character string, name of shiny input. Defaults to "_client_tz".
When used in \code{use_client_tz}, may need to be namespaced if called within a module.}

\item{session}{a \code{shiny::session} object, defaults to \code{shiny::getDefaultReactiveDomain}}
}
\value{
\code{use_client_tz()}: HTML tags to be included in Shiny UI

timezone
}
\description{
Uses some Javascript to pass the client's timezone to the Shiny session.
}
\examples{
if(interactive()){

shiny::shinyApp(
  ui = shiny::fluidPage(use_client_tz(), shiny::textOutput("time")),
  server = function(input, output, session) {
    # can be used outside of reactive context, if desired
    tz <- get_client_tz()
    time <- format(Sys.time(), format = "\%x \%X", tz = tz)
    output$time <- renderText(paste0(tz,": ", time))
  }
)
}
}
