% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str}
\alias{str}
\title{Tan's str() function}
\usage{
str(..., max.level = 2)
}
\arguments{
\item{...}{objects passed to str}

\item{max.level}{sets max.level - by default, 2}
}
\value{
output of utils::str() but defaults to max.level = 2
}
\description{
Wraps \code{utils::str()} but defaults to max.level = 2
}
\examples{

list(
  data = list(
    mtcars = data.frame(mtcars),
    airquality = data.frame(airquality)
  )
) |>
  str()
}
