% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_tantastic.R
\name{theme_tantastic}
\alias{theme_tantastic}
\title{Tan's theme - dark}
\usage{
theme_tantastic(
  base_family = "IBM Plex Sans Condensed",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = "IBM Plex Sans Condensed",
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = "IBM Plex Sans Condensed",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE
)
}
\description{
Largely inspired by hrbrthemes's modern rc
}
