% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_unbind.R
\name{unbind_dt_js}
\alias{unbind_dt_js}
\title{Unbind Shiny Inputs (UI function)}
\usage{
unbind_dt_js()
}
\value{
Adds JS to HTML head
}
\description{
When generating inputs reactively, it's sometimes necessary to unbind old inputs so that you can use the updated inputs in Shiny.
`\code{unbind_dt_js()} adds the JS to the UI, while \code{unbind_dt()} is called as needed in the server component
(usually as part of an \code{observeEvent} or \code{eventReactive})
}
\examples{
unbind_dt_js()
}
