% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str.R
\name{str}
\alias{str}
\title{Tan's str() function}
\usage{
str(...)
}
\arguments{
\item{...}{objects passed to str}
}
\value{

}
\description{
Wraps \code{utils::str()} but defaults to max.level = 2
}
\examples{

list(
  data = list(
    mtcars = mtcars,
    airquality = airquality
  )
) \%>\%
  str()
}
